+++ spec/spec_helper/user_interface.rb
@@ -1,0 +1,13 @@
+module Pod
+  class UI
+    @output = ''
+
+    class << self
+      attr_accessor :output
+
+      def puts(message = '')
+        @output << "#{message}"
+      end
+    end
+  end
+end