+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+## 0.15.0 (unreleased)
+
+[CocoaPods](https://github.com/CocoaPods/CocoaPods/compare/master...b0.15.0)
+
+###### Enhancements
+
+- Added a pre install hook to the Podfile and to root specifications.
+  [#486](https://github.com/CocoaPods/CocoaPods/issues/486)
+- Support for `header_mappings_dir` attribute in subspecs.
+- Refactored UI.
+
+###### Bug fixes
+
+- Improvements to the git cache that should be more robust.
+  [#517](https://github.com/CocoaPods/CocoaPods/issues/517)
+
 ## Master
 
 [CocoaPods](https://github.com/CocoaPods/CocoaPods/compare/0.14.0...master)